+++ ReactProtectedRoute/src/context/AuthContext.js
@@ -1,6 +1,5 @@
 import React from "react";
 
-const AuthContext = React.createContext()
+const AuthContext = React.createContext();
 
-
-export default AuthContext
+export default AuthContext;
